% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{exploreHCAexport}
\alias{exploreHCAexport}
\title{exploration app}
\usage{
exploreHCAexport()
}
\description{
exploration app
}
\note{
This shiny app, in development, will navigate the AnVIL DATA elements of an AnVIL workspace.
It is tailored at present to help explore the outputs of the Human Cell Atlas portal "Export
to terra" functions.
}
